Support ignore patterns for requests library (#15)
* Support ignore patterns for requests library

In some cases the user might want to avoid tracing of specific URL's or
endpoints. This commit adds a new parameter to the
`trace_requests()`-function to set a list of URL patterns to ignore when
a request is made. `ignore_patterns` is a keyword argument and should
contain python regex patterns that can be used for matching. For
example, ignore tracing requests to a sidecar container running in
localhost:4040:

    ```
    trace_requests(ignore_patterns=[r".*localhost:4040"])
    ```

* Fix a typo in requests library tests

* Do not attempt to remove span from headers

* Address review comments